fix(#23): 커스텀 훅 제거
멘토링 이후, 극단적인 분리는 좋지 않을 수 있다고 하여 다시 수정하였습니다!
